backend/requirements.txt
fastapi==0.115.0
azure-communication-email==1.0.0
uvicorn==0.30.0
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
python-multipart==0.0.9
pydantic==2.8.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
